SnakeTron server stuff

+++ common/src/snake.rs
@@ -1,6 +1,5 @@
 use std::collections::VecDeque;
 
-// Direction
 #[derive(Debug)]
 pub enum Direction {
     Up,
@@ -9,20 +8,19 @@ pub enum Direction {
     Right,
 }
 
-// Snake
 #[derive(Debug)]
 pub struct Snake {
-    pub body: VecDeque<(u32, u32)>,
+    pub body: VecDeque<(u16, u16)>,
     pub direction: Direction,
     pub food: u32,
 }
 
 impl Snake {
-    pub fn head(&self) -> &(u32, u32) {
+    pub fn head(&self) -> &(u16, u16) {
         self.body.front().expect("Snake body should not be empty")
     }
 
-    pub fn tail(&self) -> &(u32, u32) {
+    pub fn tail(&self) -> &(u16, u16) {
         self.body.back().expect("Snake body should not be empty")
     }
 }
